+++ configs/config.json.export.example
@@ -13,7 +13,7 @@
         "type": "CollectLevelUpReward"
       },
       {
-        "type": "ExportCsv",
+        "type": "ExportFile",
         "config": {
           "file_type": "csv"
         }
