+++ pokemongo_bot/__init__.py
@@ -460,6 +460,12 @@ class PokemonGoBot(object):
             parameters=('last_lat', 'last_lon')
         )
 
+        # Export File
+        self.event_manager.register_event(
+            'file_exported',
+            parameters=('file')
+        )
+
     def tick(self):
         self.health_record.heartbeat()
         self.cell = self.get_meta_cell()
